requirements.txt
APScheduler==3.10.4
asgiref==3.7.2
backports.zoneinfo==0.2.1
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.9
django-apscheduler==0.6.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
docker==7.0.0
idna==3.6
packaging==24.0
PyJWT==2.8.0
PyMySQL==1.1.1
pysolr==3.9.0
python-dotenv==1.0.1
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.9.0
tzlocal==5.2
urllib3==2.1.0